splitup reporters into sepearate files; splitup reporter.spec into seperate files
