+++ test/unit/reporter.spec.coffee
@@ -8,31 +8,6 @@ describe 'reporter', ->
   beforeEach ->
     m = loadFile __dirname + '/../../lib/reporter.js', {xmlbuilder: require 'xmlbuilder'}
 
-  #============================================================================
-  # Base [PRIVATE]
-  #============================================================================
-  describe 'Progress', ->
-    adapter = reporter = null
-
-    beforeEach ->
-      adapter = jasmine.createSpy 'STDOUT'
-      reporter = new m.BaseReporter null, null, adapter
-
-
-    it 'should write to all registered adapters', ->
-      anotherAdapter = jasmine.createSpy 'ADAPTER2'
-      reporter.adapters.push anotherAdapter
-
-      reporter.write 'some'
-      expect(adapter).toHaveBeenCalledWith 'some'
-      expect(anotherAdapter).toHaveBeenCalledWith 'some'
-
-
-    it 'should format', ->
-      reporter.write 'Success: %d Failure: %d', 10, 20
-
-      expect(adapter).toHaveBeenCalledWith 'Success: 10 Failure: 20'
-
 
   #==============================================================================
   # formatError() [PRIVATE]
